restore pre-VPX10.4 behavior for sounds that were re-routed to the backglass via their name/path, also make by default all the other sounds to be routed to the table for VP8/9 tables
and some leftovers, NFC

git-svn-id: https://svn.code.sf.net/p/vpinball/code/trunk@4082 14438d79-0ffd-4333-8ae7-ea2bdd901387

+++ primitive.cpp
@@ -467,7 +467,7 @@ void Primitive::GetHitShapes(vector<HitObject*> &pvho)
 
    //
 
-   const unsigned int reduced_vertices = max((unsigned int)powf((float)m_vertices.size(), clamp(1.f - m_d.m_collision_reductionFactor, 0.f, 1.f)*0.25f + 0.75f), 420u); //!! 420 = magic
+   const unsigned int reduced_vertices = max((unsigned int)pow((double)m_vertices.size(), clamp(1.f - m_d.m_collision_reductionFactor, 0.f, 1.f)*0.25f + 0.75f), 420u); //!! 420 = magic
 
    if (reduced_vertices < m_vertices.size())
    {
